+++ cli/src/commands/git/clone.rs
@@ -77,6 +77,23 @@ pub struct GitCloneArgs {
     /// while all subsequent fetches will only fetch included tags.
     #[arg(long, value_enum)]
     fetch_tags: Option<FetchTagsMode>,
+    /// Name of the branch to fetch and use as the parent of the working-copy
+    /// change
+    ///
+    /// If not present, all branches are fetched and the repository's default
+    /// branch is used as parent of the working-copy change.
+    ///
+    /// By default, the specified name matches exactly. Use `glob:` prefix to
+    /// expand `*` as a glob, e.g. `--branch 'glob:push-*'`. Other wildcard
+    /// characters such as `?` are *not* supported. Can be repeated to specify
+    /// multiple branches, in which case the first match of the first `--branch`
+    /// argument is used as the working-copy parent.
+    #[arg(
+        long, short,
+        alias = "bookmark",
+        value_parser = StringPattern::parse,
+    )]
+    branch: Option<Vec<StringPattern>>,
 }
 
 fn clone_destination_for_source(source: &str) -> Option<&str> {
@@ -145,11 +162,13 @@ pub fn cmd_git_clone(
             // If not explicitly specified on the CLI, configure the remote for only fetching
             // included tags for future fetches.
             args.fetch_tags.unwrap_or(FetchTagsMode::Included),
+            args.branch.as_deref(),
         )?;
         let default_branch = fetch_new_remote(
             ui,
             &mut workspace_command,
             remote_name,
+            args.branch.as_deref(),
             args.depth,
             args.fetch_tags,
         )?;
@@ -177,21 +196,26 @@ pub fn cmd_git_clone(
         }
     }
 
-    let (mut workspace_command, default_branch) = clone_result?;
-    if let Some(name) = &default_branch {
-        let default_symbol = name.to_remote_symbol(remote_name);
-        write_repository_level_trunk_alias(ui, workspace_command.repo_path(), default_symbol)?;
+    let (mut workspace_command, (working_branch, working_is_default)) = clone_result?;
 
-        let default_branch_remote_ref = workspace_command
+    if let Some(name) = &working_branch {
+        let working_symbol = name.to_remote_symbol(remote_name);
+        if working_is_default {
+            write_repository_level_trunk_alias(ui, workspace_command.repo_path(), working_symbol)?;
+        }
+        let working_branch_remote_ref = workspace_command
             .repo()
             .view()
-            .get_remote_bookmark(default_symbol);
-        if let Some(commit_id) = default_branch_remote_ref.target.as_normal().cloned() {
+            .get_remote_bookmark(working_symbol);
+        if let Some(commit_id) = working_branch_remote_ref.target.as_normal().cloned() {
             let mut tx = workspace_command.start_transaction();
             if let Ok(commit) = tx.repo().store().get_commit(&commit_id) {
                 tx.check_out(&commit)?;
             }
-            tx.finish(ui, "check out git remote's default branch")?;
+            tx.finish(
+                ui,
+                format!("check out git remote's branch: {}", name.as_symbol()),
+            )?;
         }
     }
 
@@ -229,6 +253,7 @@ fn configure_remote(
     remote_name: &RemoteName,
     source: &str,
     fetch_tags: FetchTagsMode,
+    branch_patterns: Option<&[StringPattern]>,
 ) -> Result<WorkspaceCommandHelper, CommandError> {
     let mut tx = workspace_command.start_transaction();
     git::add_remote(
@@ -236,6 +261,7 @@ fn configure_remote(
         remote_name,
         source,
         fetch_tags.as_fetch_tags(),
+        branch_patterns,
     )?;
     tx.finish(ui, format!("add git remote {}", remote_name.as_symbol()))?;
     // Reload workspace to apply new remote configuration to
@@ -255,9 +281,10 @@ fn fetch_new_remote(
     ui: &Ui,
     workspace_command: &mut WorkspaceCommandHelper,
     remote_name: &RemoteName,
+    target_branches: Option<&[StringPattern]>,
     depth: Option<NonZeroU32>,
     fetch_tags: Option<FetchTagsMode>,
-) -> Result<Option<RefNameBuf>, CommandError> {
+) -> Result<(Option<RefNameBuf>, bool), CommandError> {
     writeln!(
         ui.status(),
         r#"Fetching into new repo in "{}""#,
@@ -265,44 +292,84 @@ fn fetch_new_remote(
     )?;
     let settings = workspace_command.settings();
     let git_settings = settings.git_settings()?;
-    let track_default = settings.get_bool("git.track-default-bookmark-on-clone")?;
+    let should_track_default = settings.get_bool("git.track-default-bookmark-on-clone")?;
     let mut tx = workspace_command.start_transaction();
-    let mut git_fetch = GitFetch::new(tx.repo_mut(), &git_settings)?;
-    with_remote_git_callbacks(ui, |cb| {
-        git_fetch.fetch(
-            remote_name,
-            expand_fetch_refspecs(remote_name, vec![StringPattern::everything()])
-                .expect("Valid default refspec"),
-            cb,
-            depth,
-            match fetch_tags {
-                // If not explicitly specified on the CLI, override the remote
-                // configuration and fetch all tags by default since this is
-                // the Git default behavior.
-                None => Some(FetchTagsOverride::AllTags),
+    let (default_branch, import_stats) = {
+        let mut git_fetch = GitFetch::new(tx.repo_mut(), &git_settings)?;
 
-                // Technically by this point the remote should already be
-                // configured based on the CLI parameters so we shouldn't *need*
-                // to apply an override here but all the cases are expanded here
-                // for clarity.
-                Some(FetchTagsMode::All) => Some(FetchTagsOverride::AllTags),
-                Some(FetchTagsMode::None) => Some(FetchTagsOverride::NoTags),
-                Some(FetchTagsMode::Included) => None,
-            },
-        )
-    })?;
-    let default_branch = git_fetch.get_default_branch(remote_name)?;
-    let import_stats = git_fetch.import_refs()?;
-    if let Some(name) = &default_branch {
+        let fetch_refspecs = expand_fetch_refspecs(
+            remote_name,
+            target_branches
+                .unwrap_or(&[StringPattern::everything()])
+                .to_vec(),
+        )?;
+
+        with_remote_git_callbacks(ui, |cb| {
+            git_fetch.fetch(
+                remote_name,
+                fetch_refspecs,
+                cb,
+                depth,
+                match fetch_tags {
+                    // If not explicitly specified on the CLI, override the remote
+                    // configuration and fetch all tags by default since this is
+                    // the Git default behavior.
+                    None => Some(FetchTagsOverride::AllTags),
+
+                    // Technically by this point the remote should already be
+                    // configured based on the CLI parameters so we shouldn't *need*
+                    // to apply an override here but all the cases are expanded here
+                    // for clarity.
+                    Some(FetchTagsMode::All) => Some(FetchTagsOverride::AllTags),
+                    Some(FetchTagsMode::None) => Some(FetchTagsOverride::NoTags),
+                    Some(FetchTagsMode::Included) => None,
+                },
+            )
+        })?;
+
+        let import_stats = git_fetch.import_refs()?;
+
+        let default_branch = git_fetch.get_default_branch(remote_name)?;
+        (default_branch, import_stats)
+    };
+
+    let working_branch = match target_branches {
+        Some(target_branches) => {
+            // Ensure that all targets match at least one branch, and record the matching
+            // branch for the first target
+            let mut first_match = None;
+            for pattern in target_branches {
+                let mut bookmarks = tx.repo().view().remote_bookmarks(remote_name);
+                let Some((bookmark, _)) =
+                    bookmarks.find(|(bookmark, _)| pattern.is_match(bookmark.as_str()))
+                else {
+                    return Err(user_error(format!(
+                        "No such branch matching pattern: {pattern}",
+                    )));
+                };
+                if first_match.is_none() {
+                    first_match = Some(bookmark.to_owned());
+                }
+            }
+            first_match
+        }
+        None => default_branch.clone(),
+    };
+
+    let working_is_default = working_branch == default_branch;
+    if let Some(name) = &working_branch
+        && working_is_default
+        && should_track_default
+    {
+        // For convenience, create local bookmark as Git would do.
         let remote_symbol = name.to_remote_symbol(remote_name);
         let remote_ref = tx.repo().get_remote_bookmark(remote_symbol);
-        if track_default && remote_ref.is_present() {
-            // For convenience, create local bookmark as Git would do.
+        if remote_ref.is_present() {
             tx.repo_mut().track_remote_bookmark(remote_symbol);
         }
     }
     print_git_import_stats(ui, tx.repo(), &import_stats, true)?;
-    if git_settings.auto_local_bookmark && !track_default {
+    if git_settings.auto_local_bookmark && !should_track_default {
         writeln!(
             ui.hint_default(),
             "`git.track-default-bookmark-on-clone=false` has no effect if \
@@ -310,5 +377,5 @@ fn fetch_new_remote(
         )?;
     }
     tx.finish(ui, "fetch from git remote into empty repo")?;
-    Ok(default_branch)
+    Ok((working_branch, working_is_default))
 }