+++ CHANGELOG.md
@@ -66,6 +66,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   tags. These keywords may be useful in non-colocated Git repositories where
   local and exported `@git` tags can point to different revisions.
 
+*  `jj git clone` now supports the `--branch` option to specify the branch(es)
+  to fetch during clone. If present, the first matching branch is used as the
+  working-copy parent.
+
 ### Fixed bugs
 
 * `jj metaedit --author-timestamp` twice with the same value no longer